feat: Major restructure, might need to roll back lol
