+++ conf/page_info.json
@@ -46,5 +46,13 @@
             "This page shows all previously collected result files.",
             "In case of any error, collecting some more data in the 'Scrape' page"
         ]
+    },
+    "test":{
+        "route":        "/test",
+        "template":     "children/test.html",
+        "title":        "TEST",
+        "desc": [
+            "TEST PAGE"
+        ]
     }
 }
