+++ src/Retailers/funcs.py
